requirements.txt
python-telegram-bot
beautifulsoup4
pandas
requests
pytz
matplotlib